+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ### [v0.10] - 2024-10-09
 
+Minor changes:
+
+- Change boiler code to source `.bashrc` to find conda environment (instead of
+  `ml` and `activate base`). This requires `conda init` to be run beforehand.
 - Packaging: switch to `pyproject.toml` for packaging.
 
 ### [v0.9] - 2024-06-06